Bootstrap & Creating Posts
Added Bootstrap to the project. Still need to learn it more and make good use of it, but for now what I got figured out looks okay.

There is now the ability to add posts to the forum. To do this I had to patch up the dbfuncs a bit to work with the design, specifically the db_add_post function, as it's parameters didn't fit with what I needed.

Anyone can view posts but to make one a user must log in first.

There are a few bugs to be fixed but I will list those in the Issues section.