Fixing Errors, Finalizing Inbox page
Added some commments to a few of the pages, created new page for sending a msg. No db connection so it doesnt work but allows test input, as well as navigation to and from it. And fixed the issue with the errors not popping up when entering an invalid comment